+++ RELEASE_NOTES.md
@@ -1,3 +1,9 @@
+### 0.1.3-beta001
+* Added utility and timing extensions for `Promise`, `Job`, and `Alt`
+* Added `OptionJob` extensions
+* Added `ChoiceJob` extensions
+* Added `UnixTime` implementation for getting the current time in Unix seconds since the epoch
+
 ### 0.1.2
 * `SharedMap.freeze` again returns a `Job<Map<'k,'v>>`